requirements.txt
pandas==1.2.5
plotly==5.9.0
dash==2.5.1
protobuf==3.20.1
python_dateutil==2.8.2
pyarrow==8.0.0
google-cloud-appengine-logging
google-cloud-audit-log
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage
google-cloud-core
google-cloud-datastore
google-cloud-storage